==> actt-1/4-model_fitting.qmd <==
---
title: "ACTT-1: Model Fitting"
---

This file fits the Bayesian ordinal transition model using the `rmsb` package.

## Load packages / data

```{r}
# Fit the model if TRUE
# Set FALSE if the model has already been fit to save time
fit <- TRUE
```

```{r}
library(tidyverse)
library(data.table)
library(rmsb)

# Set number of cores for parallel processing
options(mc.cores = parallel::detectCores())

set.seed(777)
```

## Read in data

```{r}
df <- arrow::read_parquet("derived_data/actt1_main.parquet")
```

## Process data

```{r}
# Drop patients who have a gap in the data
df <- df[!(id %in% unique(df[gap > 1]$id))]

# Must be integer for model fitting
df$y <- as.integer(df$y)
```

## Statistical Modeling

```{r}
if (fit) {
    model <-
    blrm(
      formula = y ~ yprev + rcs(age, 4) + rcs(day, 6) * tx + sex,
      pcontrast=list(sd=2,
                     c1=list(tx='Placebo', age=50, day = 8, yprev = 4, sex = "F"),
                     c2=list(tx='Remdesivir', age=50, day = 8, yprev = 4, sex = "F"),
                     contrast=expression(c1 - c2)),
      ppo = ~ day,
      cppo = function(y) y,
      data = df,
      refresh = 5,
      iter = 2000,
      chains = 4,
      method = "sampling"
  )
  
  # Save model to an RDS file
  saveRDS(model, "fitted_models/model1.rds")
}
```
